+++ src/main/kotlin/com/nylas/models/UpdateEventRequest.kt
@@ -496,23 +496,6 @@ data class UpdateEventRequest(
   }
 
   /**
-   * Class representing the reminders field of an event.
-   */
-  data class Reminders(
-    /**
-     * Whether to use the default reminders for the calendar.
-     * When true, uses the default reminder settings for the calendar
-     */
-    @Json(name = "use_default")
-    val useDefault: Boolean? = null,
-    /**
-     * A list of reminders for the event if useDefault is set to false.
-     */
-    @Json(name = "override")
-    val override: List<ReminderOverride>? = null,
-  )
-
-  /**
    * Builder for [UpdateEventRequest].
    */
   class Builder {